feat: vitest plugin
