+++ examples/js/vitest.config.ts
@@ -1,3 +1,7 @@
 import { defineConfig } from "vitest/config";
+import { ddtPlugin } from "@ddtds/vitest";
 
-export default defineConfig({ test: { environment: "node" } });
+export default defineConfig({
+  plugins: [ddtPlugin(".", "__doctests__")],
+  test: { environment: "node" },
+});
