+++ examples/react/vitest.config.ts
@@ -1,6 +1,8 @@
 import { defineConfig } from "vitest/config";
+import { ddtPlugin } from "@ddtds/vitest";
 
 export default defineConfig({
+  plugins: [ddtPlugin(".", "__doctests__")],
   test: {
     environment: "jsdom",
     setupFiles: ["@testing-library/jest-dom/vitest"],
