+++ packages/core/package.json
@@ -14,6 +14,7 @@
     "unist-util-visit": "^5.1.0"
   },
   "devDependencies": {
+    "@types/node": "catalog:",
     "vitest": "catalog:"
   }
 }
