+++ packages/vitest/package.json
@@ -13,6 +13,7 @@
     "@ddtds/core": "workspace:*"
   },
   "devDependencies": {
+    "vite": "^8.0.10",
     "vitest": "catalog:"
   }
 }
